refactor: se ha modificado el serviceImplements de empleado

+++ Proyecto/src/main/java/com/JosueJolon/RepuestosAutomotricez/Service/EmpleadoServiceImplements.java
@@ -1,8 +1,10 @@
 package com.JosueJolon.RepuestosAutomotricez.Service;
 
 
-import com.JosueJolon.RepuestosAutomotricez.Models.Empleado;
+import com.JosueJolon.RepuestosAutomotricez.Entity.Empleado;
+import com.JosueJolon.RepuestosAutomotricez.Exception.ResourceNotFoundException;
 import com.JosueJolon.RepuestosAutomotricez.Repository.EmpleadoRepository;
+import com.JosueJolon.RepuestosAutomotricez.Validator.EmpleadosValidator;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -10,12 +12,12 @@ import java.util.List;
 @Service
 public class EmpleadoServiceImplements implements EmpleadoService {
     private final EmpleadoRepository empleadoRepository;
+    private final EmpleadosValidator empleadosValidator;
 
-    public EmpleadoServiceImplements(EmpleadoRepository empleadoRepository){
+    public EmpleadoServiceImplements(EmpleadoRepository empleadoRepository, EmpleadosValidator empleadosValidator) {
         this.empleadoRepository = empleadoRepository;
+        this.empleadosValidator = empleadosValidator;
     }
-
-
 
     @Override
     public List<Empleado> getAListEmpleados() {
@@ -24,11 +26,17 @@ public class EmpleadoServiceImplements implements EmpleadoService {
 
     @Override
     public Empleado getEmpleadoById(Integer id) {
-        return empleadoRepository.findById(id).orElse(null);
+        Empleado empleado = empleadoRepository.findById(id).orElse(null);
+        if(empleado == null){
+            throw new IllegalArgumentException();
+        }else {
+            return empleadoRepository.findById(id).orElse(null);
+        }
     }
 
     @Override
     public Empleado saveEmpleado(Empleado empleado) throws RuntimeException {
+        empleadosValidator.validarEmpleado(empleado);
         return empleadoRepository.save(empleado);
     }
 
@@ -36,18 +44,24 @@ public class EmpleadoServiceImplements implements EmpleadoService {
     public Empleado updateEmpleado(Integer id, Empleado empleado) {
         Empleado empleados = empleadoRepository.findById(id).orElse(null);
         if(empleados != null){
+            empleadosValidator.validarEmpleado(empleado);
             empleados.setNombre_empleado(empleado.getNombre_empleado());
             empleados.setApellido_empleado(empleado.getApellido_empleado());
             empleados.setPuesto_empleado(empleado.getPuesto_empleado());
             empleados.setEmail_empleado(empleado.getEmail_empleado());
         } else {
-            throw new IllegalArgumentException();
+            throw new ResourceNotFoundException("el id del empleado no existe");
         }
         return empleadoRepository.save(empleados);
     }
 
     @Override
     public void deleteEmpleado(Integer id) {
+        Empleado empleado = empleadoRepository.findById(id).orElse(null);
+        if(empleado == null){
+            throw new ResourceNotFoundException("el id del empleado no existe");
+        }
+
         empleadoRepository.deleteById(id);
     }
 }
